+++ MagicAPI/src/main/java/com/elmakers/mine/bukkit/api/action/SpellAction.java
@@ -25,8 +25,20 @@ public interface SpellAction extends Cloneable
     boolean requiresTargetEntity();
     boolean ignoreResult();
     String transformMessage(String message);
-    void load(Mage mage, ConfigurationSection data);
-    void save(Mage mage, ConfigurationSection data);
     int getActionCount();
     Object clone();
+
+    /**
+     * This mechanism never worked properly and is no longer called.
+     * Actions that need to store data should interact with CastContext.getData instead.
+     */
+    @Deprecated
+    default void load(Mage mage, ConfigurationSection data) { }
+
+    /**
+     * This mechanism never worked properly and is no longer called.
+     * Actions that need to store data should interact with CastContext.setData instead.
+     */
+    @Deprecated
+    default void save(Mage mage, ConfigurationSection data) { }
 }